bigdata-report/requirements.txt
JPype1==1.3.0
JayDeBeApi==1.2.3
loguru==0.5.3
Flask==1.1.4
aiohttp==3.7.4.post0
aiomysql==0.0.21